Fix beta app re-render error and invest success panel state (#84)
When an invest or redeem transaction completed, and you click on the `invest more` or `redeem more` button to return to initial form inputs component, the invest/redeem page would seem to disappear. This was caused by an infinite loop of re-renders coming from our `useObservable` hook. The hook would continuously call `onStoreChange`. A more robust setup was configured to help prevent this. The invest success panel was also updated to reflect the current state of the transaction as it is signed, pending, and either successful or failed.